requirements.txt
asgiref==3.4.1
biopython==1.79
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-import-export==2.7.1
et-xmlfile==1.1.0
MarkupPy==1.14
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.10
Pillow==8.4.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.1.0
typing-extensions==4.1.1
xlrd==2.0.1
xlwt==1.3.0
